use Float more consistently

+++ src/config.rs
@@ -4,4 +4,4 @@ pub type Float = f64;
 pub type Color = ColorRgb;
 pub type Film = SamplingFilm;
 
-pub const PI: f64 = std::f64::consts::PI;
+pub const PI: Float = std::f64::consts::PI;
